requirements.txt
backports.shutil-get-terminal-size==1.0.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
configparser==3.5.0
enum34==1.1.6
ExifRead==2.1.2
flake8==3.4.1
Flask==1.0.2
gunicorn==19.9.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
numpy==1.15.0
pathlib==1.0.1
pew==1.1.0
Pillow==5.2.0
pipenv==8.2.7
psutil==5.3.1
pycodestyle==2.3.1
pyflakes==1.5.0
requests==2.18.4
requirements==0.1
shutilwhich==1.1.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
Werkzeug==0.14.1
